middleware/requirements.txt
appnope==0.1.3
asttokens==2.2.1
azure-cognitiveservices-speech==1.24.2
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
docker==6.0.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==5.2.0
ipykernel==6.19.4
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
torch==1.13.1
tornado==6.2
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.11.0
